+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderFacade.java
@@ -1,20 +1,11 @@
 package com.loopers.application.order;
 
-import com.loopers.application.payment.PaymentCriteria;
-import com.loopers.application.payment.PaymentFacade;
 import com.loopers.domain.coupon.CouponEntity;
 import com.loopers.domain.coupon.CouponService;
 import com.loopers.domain.order.*;
 import com.loopers.domain.order.event.OrderCreatedEvent;
 import com.loopers.domain.payment.PaymentMethod;
-import com.loopers.domain.payment.PaymentRepository;
-import com.loopers.domain.payment.PaymentService;
-import com.loopers.domain.point.PointService;
 import com.loopers.domain.product.ProductService;
-import com.loopers.domain.userCoupon.UserCouponCommand;
-import com.loopers.domain.userCoupon.UserCouponService;
-import com.loopers.support.error.CoreException;
-import com.loopers.support.error.ErrorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
@@ -28,7 +19,6 @@ public class OrderFacade {
 
     private final OrderService orderService;
     private final CouponService couponService;
-    //private final PaymentFacade paymentFacade;
     private final ProductService productService;
     private final ApplicationEventPublisher publisher;
 
@@ -54,17 +44,6 @@ public class OrderFacade {
         }
         long payable = Math.max(0, totalPrice - discount);
 
-        /* 결제 요청
-        paymentFacade.requestPayment(new PaymentCriteria.RequestPayment(
-                saved.getId(),
-                c.userId(),
-                payable,
-                PaymentMethod.CARD,
-                new PaymentCriteria.RequestPayment.Card(c.cardType(), c.cardNo()),
-                c.couponId()
-        ));
-        */
-        // 커밋 이후 비동기로 후속 처리하도록 이벤트 발행
         publisher.publishEvent(new OrderCreatedEvent(
                 saved.getId(),
                 c.userId(),
